refactor(server): separate app construction from runtime startup
